Рефакторит код адаптера
