+++ app/src/main/java/ru/practicum/android/diploma/common/di/DataModule.kt
@@ -1,5 +1,6 @@
 package ru.practicum.android.diploma.common.di
 
+import android.app.Application
 import android.content.Context
 import androidx.room.Room
 import com.google.gson.Gson
@@ -18,6 +19,8 @@ import ru.practicum.android.diploma.search.domain.api.FavoritesDBConverter
 import ru.practicum.android.diploma.search.domain.api.SearchResultConverter
 import ru.practicum.android.diploma.search.domain.impl.FavoritesDBConverterImpl
 import ru.practicum.android.diploma.search.domain.impl.SearchResultConverterImpl
+import ru.practicum.android.diploma.vacancy.data.sharing.ExternalNavigatorImpl
+import ru.practicum.android.diploma.vacancy.domain.api.ExternalNavigator
 import ru.practicum.android.diploma.vacancy.domain.api.SingleVacancyConverter
 import ru.practicum.android.diploma.vacancy.domain.impl.SingleVacancyConverterImpl
 import javax.inject.Singleton
@@ -57,4 +60,7 @@ class DataModule {
 
     @Provides
     fun providesSingleVacancyConverter(): SingleVacancyConverter = SingleVacancyConverterImpl(gson = provideGson())
+
+    @Provides
+    fun providesExternalNavigator(): ExternalNavigator = ExternalNavigatorImpl(application = Application())
 }
