+++ app/src/main/java/ru/practicum/android/diploma/common/di/InteractorModule.kt
@@ -12,6 +12,7 @@ import ru.practicum.android.diploma.search.domain.api.SearchInteractor
 import ru.practicum.android.diploma.search.domain.api.SearchResultConverter
 import ru.practicum.android.diploma.search.domain.impl.FavoritesDBInteractorImpl
 import ru.practicum.android.diploma.search.domain.impl.SearchInteractorImpl
+import ru.practicum.android.diploma.vacancy.domain.api.ExternalNavigator
 import ru.practicum.android.diploma.vacancy.domain.api.SingleVacancyConverter
 import ru.practicum.android.diploma.vacancy.domain.api.SingleVacancyInteractor
 import ru.practicum.android.diploma.vacancy.domain.impl.SingleVacancyInteractorImpl
@@ -40,6 +41,7 @@ class InteractorModule {
     fun providesSingleVacancyInteractor(
         repository: HHSearchRepository,
         converter: SingleVacancyConverter,
-        database: AppDatabase
-    ): SingleVacancyInteractor = SingleVacancyInteractorImpl(repository, converter, database)
+        database: AppDatabase,
+        externalNavigator: ExternalNavigator
+    ): SingleVacancyInteractor = SingleVacancyInteractorImpl(repository, converter, database, externalNavigator)
 }
