+++ app/src/main/java/ru/practicum/android/diploma/root/ui/RootActivity.kt
@@ -40,7 +40,7 @@ class RootActivity : BaseActivity<ActivityRootBinding>(ActivityRootBinding::infl
 
         navController.addOnDestinationChangedListener { _, destination, _ ->
             when (destination.id) {
-                R.id.filterFragment -> bottomNavigationView.visibility = View.GONE
+                R.id.filterFragment, R.id.vacancyFragment -> bottomNavigationView.visibility = View.GONE
                 else -> bottomNavigationView.visibility = View.VISIBLE
             }
         }
